refactor: move signal code in model to separate file
