+++ user/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class UserConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "user"
+
+    def ready(self):
+        from . import signals
