+++ user/models.py
@@ -1,11 +1,8 @@
-from django.db import models
 from django.conf import settings
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-from django.contrib.auth.models import User
+from django.db import models
 from django.urls import reverse
 
-# Create your models here.
+
 class Profile(models.Model):
     user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     bio = models.TextField(max_length=500, blank=True)
@@ -18,14 +15,3 @@ class Profile(models.Model):
 
     def get_absolute_url(self):
         return reverse("user:profile", args=(self.id,))
-
-
-@receiver(post_save, sender=User)
-def create_user_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
-
-
-@receiver(post_save, sender=User)
-def save_user_profile(sender, instance, **kwargs):
-    instance.profile.save()
